Correction to follow stck architecture
