+++ lib/ui/view/workspace/email confirmation view.dart
@@ -1,29 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:hexcolor/hexcolor.dart';
-import 'dart:io'; 
-// ignore: import_of_legacy_library_into_null_safe
-import 'package:flutter_appavailability/flutter_appavailability.dart';
+// ignore: unused_import
+import 'email confirmation model.dart';
+
 void main() => runApp(MyApp());
+final EmailViewModel emailViewModel = new EmailViewModel();
 
 class MyApp extends StatelessWidget {
-
- void openEmailApp(BuildContext context){
-     try{
-         AppAvailability.launchApp(Platform.isIOS ? "message://" : "com.google.android.gm").then((_) {
-                 print("App Email launched!");
-               }).catchError((err) {
-                 // ignore: deprecated_member_use
-                 Scaffold.of(context).showSnackBar(SnackBar(
-                     content: Text("App Email not found!")
-                 ));
-                 print(err);
-               });
-     } catch(e) {
-       // ignore: deprecated_member_use
-       Scaffold.of(context).showSnackBar(SnackBar(content: Text("Email App not found!")));
-     }
- }
-
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -46,7 +29,7 @@ class MyApp extends StatelessWidget {
                     height: 10,
                   ),
                   Text(
-                      "To confirm your email address, tap the button in the email we sent to lukeskylwalker@gmai.com",
+                      "To confirm your email address, tap the button in the email we sent to lukeskylwalker@gmail.com",
                       textAlign: TextAlign.center,
                       style: TextStyle(
                           fontFamily: "Lato",
@@ -68,7 +51,8 @@ class MyApp extends StatelessWidget {
                     ),
                     onPressed: () {
                       print('Pressed');
-                      openEmailApp(context);
+                      emailViewModel.openEmailApp(context);
+                      // openEmailApp(context);
                     },
                   )
                 ]),
